SystemMonitor: Use u64 instead of size_t in page_count_to_bytes function
Apparently with huge amount of memory (I forced the kernel to think we
have about 282 GiB of memory), a size_t count is not enough and instead
a u64 type wrap it better.

+++ Userland/Applications/SystemMonitor/MemoryStatsWidget.cpp
@@ -62,7 +62,7 @@ MemoryStatsWidget::~MemoryStatsWidget()
 {
 }
 
-static inline u64 page_count_to_bytes(size_t count)
+static inline u64 page_count_to_bytes(u64 count)
 {
     return count * 4096;
 }
